+++ price_service/server/src/helpers.ts
@@ -34,6 +34,8 @@ export function removeLeading0x(s: string): string {
   return s;
 }
 
+// Helper for treating T | undefined as an optional value. This lets you pick a
+// default if value is undefined.
 export function getOrElse<T>(value: T | undefined, defaultValue: T): T {
   if (value === undefined) {
     return defaultValue;
